pgw#1362 wave 2: the cli/config/sdk cluster folds from 9 incident files into 3 domain modules
  9 modules / 2,660 lines  ->  3 modules / 2,226 lines   (-67% files, -16% lines)

    test_config_authority.py  the declared interpreter env, and the torch flags it imposes
    test_config_boundary.py   who may read settings, and how a reconcile resolves
    test_sdk_authoring.py     the decorators, the job contract, what a handler emits

plus test_model_sdk_pgw1332.py -> test_model_sdk.py: 1,320 lines, already single-domain and
already torch-gated. It needed a NAME, not a merge, and forcing it into one would have bought
nothing but a worse file.

ALL 120 TESTS SURVIVE, counted on both trees before the originals were deleted.

THE TORCH GATE DID NOT WIDEN, and this is the trap wave 1 named. A module-level
`pytest.importorskip("torch")` poisons a merge: it turns a per-module skip into a skip of every
neighbour that never needed torch, silently, on any runner missing the extra. So
test_settings_authority_pgw1049 and test_env_contract_pgw718 — both already gated, both torch in
every test — merged with EACH OTHER, and the three torchless config modules merged separately.

ONE REDUNDANT FIXTURE DELETED RATHER THAN CARRIED. pgw#718's `_reset_boot_seal` unwound
`env_seal._BOOT_READBACK`; pgw#1049's `_fresh_boot`, now module-wide, already unwinds that plus
`_ESTABLISHED_OVERRIDES`. Only reset-shaped autouse fixtures may widen on a merge — they can only
add isolation — and a strictly-subsumed one is dead weight.

FOUR MORE MODULES OFF THE MYPY BURN-DOWN, none added: high-water 255 -> 251 on this branch.
`mypy src tests` green over 937 files; all 24 fast-gate lint scripts pass locally.

MUTATION-PROBED, one load-bearing guard per module:

  config_authority  settings_authority.impose_torch stops imposing cudnn_benchmark   2 RED
  config_boundary   runtime_config.observe stops writing the snapshot                2 RED
  sdk_authoring     api/jobs stops refusing a class-declared @job                     1 RED

Three prose cross-references that named folded files were updated with them:
`api/jobs.py`, `scripts/lint_serving_process_compiles.py`, `tests/test_job_dispatch_pgw1324.py`.

Three modules in this cluster were deliberately NOT folded: test_import_cycles_pgw981,
test_retired_env_names_pgw1305 and test_cozy_runtime_env_contract_pgw1237 are GUARDS (the last a
corpus-digest gate), and the epic's brief leaves the guard family alone.
