+++ pyproject.toml
@@ -448,9 +448,7 @@ module = [
   "tests.test_component_vocab_sweep_pgw740",
   "tests.test_compute_dtype_sweep_pgw1019",
   "tests.test_compute_dtype_uniformity_pgw683",
-  "tests.test_config_boundary_pgw931",
   "tests.test_config_declaration_th1087",
-  "tests.test_config_reconcile_th1087",
   "tests.test_content_credentials_pgw658",
   "tests.test_control_plane_keepalive_pgw1125",
   "tests.test_countable_container_formula_pgw1018",
@@ -473,7 +471,6 @@ module = [
   "tests.test_dynamo_no_publish_pgw1010",
   "tests.test_eager_first_boot_pgw671",
   "tests.test_engine_matrix_pgw1011",
-  "tests.test_env_contract_pgw718",
   "tests.test_env_seal_boot_pgw694",
   "tests.test_error_visibility_pgw760",
   "tests.test_escape_hatch_pgw1062",
@@ -587,7 +584,6 @@ module = [
   "tests.test_rigboot_pgw1120",
   "tests.test_rigcheck_pgw1114",
   "tests.test_scope_armed_slot_target_pgw1078",
-  "tests.test_sdk_v2_pgw647",
   "tests.test_seal_lib_memo_pgw832",
   "tests.test_serve_finalize_pgw672",
   "tests.test_shape_growth_pgw916",
